Adding changes of decrypt button
